Create string library
This first commit has two main functions: append and split. These aims
to copy the behaviour of the two Python functions of same name.

Append works only with arrays of strings, by expanding its size with
realloc and setting its last value with the desired string.

Split gets a string and splits it into how many needed strings every
time it finds a splitting char, also given to it as a parameter. There
is also need for a parameter of the string total size and a pointer to
an int that will store the final array size.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>